Remove controller trait
