+++ plt/plt-scheduler/src/protocol_level_locks.rs
@@ -1,6 +1,5 @@
 pub mod lock_configuration;
 pub mod lock_controller;
-mod lock_controller_simple;
 
 pub mod p11;
 pub mod p9;
